+++ R/fill.R
@@ -54,5 +54,5 @@ fill_.data.frame <- function(data, fill_cols, .direction = c("down", "up")) {
 
 #' @export
 fill_.grouped_df <- function(data, fill_cols, .direction = c("down", "up")) {
-  do(data, fill_(., fill_cols = fill_cols, .direction = .direction))
+  dplyr::do(data, fill_(., fill_cols = fill_cols, .direction = .direction))
 }